pruebaluigui/requirements.txt
Django==1.11.4
djangorestframework==3.6.2
django-rest-auth
django-allauth
psycopg2